requirements.txt
langchain-openai
openai
python-dotenv
opencv-python
moviepy
pytube